refactor(frontend): decompose NetworkDetailPage god-component
Split the 938-line NetworkDetailPage into a ~370-line orchestrator plus
cohesive units, behaviour-preserving:

- hooks/useNetworkDetailData.ts — all server state, the parallel initial
  load, background polling, and the 14 mutation handlers.
- hooks/useChangeEventFilters.ts — severity/type/reviewed filtering +
  pagination for the change-event list (self-contained to that card).
- components/MonitorInfoCard.tsx — the collapsible "how it works" explainer
  (was an inline component).
- components/SectionSideNav.tsx — the sticky IntersectionObserver section nav
  (was an inline component) + the SectionDef type.
- components/ChangeEventsSection.tsx — the Change Events card (owns the filter
  hook internally).
- components/HelpPopover.tsx — extracts the 4 repeated info-popover blocks.

The page now wires the data hook to presentation and owns only the
add-snapshot modal toggle + nav-section list. Public export unchanged
(router imports the named NetworkDetailPage).

Verified: tsc --noEmit clean, docker compose build, Playwright walkthrough of
the live monitor network — all 11 sections, 9 side-nav links, info-card
expand, change-event severity filter, and help popovers render with no
console errors.
